requirements.txt
asgiref==3.7.2
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
cryptography==45.0.2
daphne==4.0.0
defusedxml==0.7.1
django==4.2.11
django-debug-toolbar==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
drf-nested-routers==0.94.2
drf-spectacular==0.26.5
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
msgpack==1.1.0
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.7.0
pymysql==1.1.1
pyopenssl==25.1.0
python-dotenv==1.1.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.2
redis==6.1.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.25.1
service-identity==24.2.0
setuptools==80.8.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.4.4
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.23.2
zope-interface==7.2
